backend/requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
chromadb==0.4.15
groq==0.4.1
sentence-transformers==2.2.2
tiktoken==0.5.1
PyPDF2==3.0.1
python-docx==1.1.0
torch==2.1.0
transformers==4.35.0
numpy==1.24.3
scipy==1.11.4